Refactor utils test suite into utils/tests/
